Fix fish calculations

+++ sisyphish/Sisyphish/Models/Fisher.cs
@@ -8,6 +8,6 @@ public class Fisher
     [FirestoreProperty("created_at")] public DateTime? CreatedAt { get; set; }
     [FirestoreProperty("discord_user_id")] public string? DiscordUserId { get; set; }
     [FirestoreProperty("fish_caught")] public List<Dictionary<string, object>> Fish { get; set; } = new();
-    public long? FishCaught => Fish.Count();
-    public long? BiggestFish => Fish.Max(f => (long)f["size"]);
+    public int? FishCaught => Fish.Count();
+    public int? BiggestFish => Fish.Max(f => (int)f["size"]);
 }
